+++ bgameb/shakers.py
@@ -4,6 +4,7 @@ from typing import Dict, Tuple, TypeVar, List, Union
 from dataclasses import dataclass, field
 from dataclasses_json import DataClassJsonMixin
 from bgameb.rollers import BaseRoller
+from bgameb.utils import log_me
 
 
 RollerCls = TypeVar('RollerCls', bound=BaseRoller)
@@ -16,14 +17,22 @@ class Shaker(DataClassJsonMixin):
     """Create shaker for roll dices or flip coins
     """
     name: str = 'shaker'
-    rollers: ShakerRollers = field(default_factory=dict, init=False) #TODO: use namedtuple or dataclass
-    last_roll: ShakerResult = field(default_factory=dict, init=False) #TODO: use namedtuple or dataclass
+    # TODO: use namedtuple or dataclass
+    rollers: ShakerRollers = field(default_factory=dict, init=False)
+    last: ShakerResult = field(default_factory=dict, init=False)
 
-    def __post_init__(self):
+    def __post_init__(self) -> None:
         self.rollers = {}
-        self.last_roll = {}
+        self.last = {}
 
-    def add(self, roller: RollerCls, color: str = 'white', count: int = 1) -> None:
+        # set logger
+        self.logger = log_me.bind(
+            classname=self.__class__.__name__,
+            name=self.name)
+        self.logger.info(f'Shaker created.')
+
+    def add(self, roller: RollerCls, color: str = 'white',
+            count: int = 1) -> None:
         """Add roller to shaker
 
         Args:
@@ -32,21 +41,41 @@ class Shaker(DataClassJsonMixin):
             count (int, optional): count of rollers copy. Defaults to 1.
 
         Raises:
-            RollerDefineError: counts not set or are different rollers with same names
+            RollerDefineError: counts not set or are different
+                               rollers with same names
         """
         if count < 1:
+            self.logger.debug(f"Can't add 0 rollers.")
             raise RollerDefineError('Need at least one roller')
+
         if self.rollers.get(color):
+
             if roller.name not in self.rollers[color].keys():
-                self.rollers[color][roller.name] = {'roller': roller, 'count': count}
+                self.rollers[color][roller.name] = {
+                    'roller': roller, 'count': count
+                    }
+
             elif self.rollers[color][roller.name]['roller'] is not roller:
-                raise RollerDefineError(
-                    f'Different instances of roller class with same name {roller.name}'
+                self.logger.debug(
+                    "Can't add different rollers with the same name."
                     )
+                raise RollerDefineError(
+                    'Different instances of roller class '
+                    f'with the same name {roller.name}'
+                    )
+
             else:
                 self.rollers[color][roller.name]['count'] += count
+                self.logger.debug(f'Number of rollers named' +
+                                  f'"{roller.name}" increased by {count}')
+
         else:
-            self.rollers[color] = {roller.name: {'roller': roller, 'count': count}}
+            self.rollers[color] = {
+                roller.name: {'roller': roller, 'count': count}
+                }
+            self.logger.debug(
+                f'Added {count} rollers with named "{roller.name}"'
+                )
 
     def remove(self, name: str, count: int, color: str) -> None:
         """Remove any kind of roller copy from shaker by name and color
@@ -57,23 +86,42 @@ class Shaker(DataClassJsonMixin):
             color (str): color froup of rollers
 
         Raises:
-            RollerDefineError: name, color not match or count of rollers not defined
+            RollerDefineError: name, color not match or
+                               count of rollers not defined
         """
         if count < 1:
+            self.logger.debug(f"Can't remove 0 rollers.")
             raise RollerDefineError('Need at least one roller')
+
         if self.rollers.get(color):
+
             if self.rollers[color].get(name):
+
                 if self.rollers[color][name]['count'] <= count:
                     del self.rollers[color][name]
+                    self.logger.debug(
+                        f'Removed rollers with {color=} and {name=}'
+                        )
+
                     if len(self.rollers[color]) == 0:
                         del self.rollers[color]
+                        self.logger.debug(
+                            f'Removed empty {color=} from shaker'
+                            )
+
                 else:
                     self.rollers[color][name]['count'] -= count
+                    self.logger.debug(f'Number of rollers named "{name}" ' +
+                                      f'with {color=} decreased by {count}')
+
             else:
+                self.logger.debug(f"Cant find roller with name {name}.")
                 raise RollerDefineError(
                     f'Cant find roller with name {name}'
                     )
+
         else:
+            self.logger.debug(f"Cant find roller with collor {color}.")
             raise RollerDefineError(
                 f'Cant find roller with collor {color}'
                 )
@@ -87,6 +135,7 @@ class Shaker(DataClassJsonMixin):
         """
         if color in self.rollers.keys():
             del self.rollers[color]
+            self.logger.debug(f'Removed rollers with {color=}')
 
     def remove_all_by_name(self, name: str) -> None:
         """Remove all rollers by roller name from shaker.
@@ -98,8 +147,12 @@ class Shaker(DataClassJsonMixin):
         to_del = []
         for color in self.rollers.keys():
             self.rollers[color].pop(name, None)
+
             if len(self.rollers[color]) == 0:
                 to_del.append(color)
+
+        self.logger.debug(f'Removed rollers with {name=}')
+
         if to_del:
             self._remove_empty_colors(to_del)
 
@@ -107,6 +160,7 @@ class Shaker(DataClassJsonMixin):
         """Remove all rollers from shaker
         """
         self.rollers = {}
+        self.logger.debug('Removed all rollers from shaker')
 
     def _remove_empty_colors(self, to_del: List[str]) -> None:
         """Remove empty colors items from rollers dict
@@ -116,6 +170,7 @@ class Shaker(DataClassJsonMixin):
         """
         for color in to_del:
             del self.rollers[color]
+            self.logger.debug(f'Removed empty {color=} from shaker')
 
     def roll(self) -> ShakerResult:
         """Roll all rollers with shaker and return results
@@ -136,13 +191,14 @@ class Shaker(DataClassJsonMixin):
                     obj['roller'].roll() for _ in range(obj['count'])
                     )
         if roll:
-            self.last_roll = roll
-            return self.last_roll
+            self.last = roll
+            self.logger.debug(f'Rolled: {roll}')
+            return self.last
         else:
+            self.logger.debug(f'No one roller rolled.')
             return {}
 
 
 class RollerDefineError(AttributeError):
     """Count of rollers not defined
     """
-    pass